2b) dino_player.h et dino_player.cpp
